+++ tests/test_simulation.py
@@ -5,43 +5,42 @@
 import numpy as np
 import pytest
 
-from pyssa.simulation import Simulation
+from cayenne.simulation import Simulation
 
 
-@pytest.mark.parametrize(
-    "algorithm", ["direct", "tau_leaping", "tau_adaptive"]
-)
+@pytest.mark.parametrize("algorithm", ["direct", "tau_leaping", "tau_adaptive"])
 @pytest.mark.usefixtures("setup_basic", "setup_large")
 class TestSanitizeAlg:
     """
         Sanity checks on Simulation class where simulations are attempted.
     """
+
     def test_null(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         k = np.array([0.0, 0.0])
-        sim = Simulation(V_r, V_p, X0, k)
+        sim = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim.simulate(algorithm=algorithm)
         assert sim.results.status_list[0] == -2
 
     def test_status_3(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         V_p = np.array([[0, 0], [0, 0], [0, 1]])
         X0 = np.array([10, 0, 0], dtype=np.int64)
-        sim = Simulation(V_r, V_p, X0, k)
+        sim = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim.simulate(algorithm=algorithm, max_t=10, max_iter=100, chem_flag=True)
         assert sim.results.status_list[0] == 3
 
     def test_status_2(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         X0 = np.array([10, 0, 0], dtype=np.int64)
-        sim = Simulation(V_r, V_p, X0, k)
+        sim = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim.simulate(algorithm=algorithm, max_t=1, max_iter=100, chem_flag=True)
         assert sim.results.status_list[0] == 2
 
     def test_reproduce(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
-        sim1 = Simulation(V_r, V_p, X0, k)
-        sim2 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
+        sim2 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim1.simulate(algorithm=algorithm)
         sim2.simulate(algorithm=algorithm)
         assert all(
@@ -55,9 +54,9 @@ class TestSanitizeAlg:
         )
 
     def test_reproduce_fail(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
-        sim1 = Simulation(V_r, V_p, X0, k)
-        sim2 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
+        sim2 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim1.simulate(algorithm=algorithm)
         sim2.simulate(algorithm=algorithm, seed=1)
         for i, j in zip(sim1.results.t_list, sim2.results.t_list):
@@ -70,30 +69,30 @@ class TestSanitizeAlg:
                 )
 
     def test_incorrect_seed(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
-        sim1 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         with pytest.raises(TypeError):
             sim1.simulate(algorithm=algorithm, n_rep=2, seed=[1])
         with pytest.raises(TypeError):
             sim1.simulate(algorithm=algorithm, n_rep=2, seed=[1, 2, 3])
 
     def test_maxiter_type(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
-        sim1 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         with pytest.raises(TypeError):
             sim1.simulate(algorithm=algorithm, max_iter=100.0)
 
     def test_monotonic(self, algorithm, setup_basic):
-        V_r, V_p, X0, k = setup_basic
-        sim1 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim1.simulate(algorithm=algorithm)
         results = sim1.results
         for t_array in results.t_list:
             assert (np.diff(t_array) > 0).all()
 
     def test_long(self, algorithm, setup_long):
-        V_r, V_p, X0, k = setup_long
-        sim1 = Simulation(V_r, V_p, X0, k)
+        species_names, rxn_names, V_r, V_p, X0, k = setup_long
+        sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         sim1.simulate(
             algorithm=algorithm, max_t=1e5, max_iter=int(1e8), chem_flag=False
         )
@@ -106,73 +105,74 @@ class TestSanitize:
     """
         Sanity checks on Simulation class where instance creation fails.
     """
+
     def test_too_high_order(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         V_r = np.array([[2, 0], [2, 1], [0, 0]])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_neg_k(self, setup_large):
-        V_r, V_p, X0, k = setup_large
+        species_names, rxn_names, V_r, V_p, X0, k = setup_large
         k = np.array([1, 1, -1, 1, -1])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_k_shape(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         k = np.array([[1], [0], [0]])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_Vp_Vr_shape(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         V_p = np.array([[0], [1], [0]])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_kdet_Vr_shape(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         k = np.array([1, 1, 1])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_X0_Vr_shape(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         X0 = np.array([100, 0, 0, 0])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         X0 = np.array([100, 0])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_X0_2d(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         X0 = np.array([[100, 0, 0]])
         print(X0.shape)
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
         X0 = np.array([[100], [0], [0]])
         print(X0.shape)
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_Vp_neg(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         V_p = np.array([[0, 0], [-1, 0], [0, 1]])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_Vr_neg(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         V_r = np.array([[-1, 0], [0, 1], [0, 0]])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
     def test_X0_neg(self, setup_basic):
-        V_r, V_p, X0, k = setup_basic
+        species_names, rxn_names, V_r, V_p, X0, k = setup_basic
         X0 = np.array([-10, 0, 0])
         with pytest.raises(ValueError):
-            Simulation(V_r, V_p, X0, k)
+            Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 
 
 # def test_bifurcation(setup_bifurcation):
@@ -180,7 +180,7 @@ class TestSanitize:
 #     count_excitation = 0
 #     n_runs = 10
 #     deviation_tolerance = 0.05
-#     sim1 = Simulation(V_r, V_p, X0, k)
+#     sim1 = Simulation(species_names, rxn_names, V_r, V_p, X0, k)
 #     sim1.simulate(
 #         algorithm="tau_adaptive", max_t=150, max_iter=1000, chem_flag=True, n_rep=n_runs
 #     )
@@ -199,12 +199,15 @@ class TestHOR:
     """
         Test the HOR method for various combinations of reactants.
     """
+
     @staticmethod
     def create_sim_inst(react_stoic):
         prod_stoic = np.random.randint(0, 5, react_stoic.shape)
         k = np.random.rand(react_stoic.shape[1])
         X = np.random.randint(0, 5, react_stoic.shape[0], dtype=np.int64)
-        sim = Simulation(react_stoic, prod_stoic, X, k)
+        species_names = ["X"] * react_stoic.shape[0]
+        rxn_names = ["R"] * react_stoic.shape[1]
+        sim = Simulation(species_names, rxn_names, react_stoic, prod_stoic, X, k)
         return sim
 
     def test_hor1(self):
@@ -293,3 +296,17 @@ class TestHOR:
         )
         sim = self.create_sim_inst(react_stoic)
         assert np.all(sim.HOR == [-3, -2, 3, 3, 3])
+
+
+class TestLoadModel:
+    def test_correct_model_str(self, setup_00001_correct):
+        sim = Simulation.load_model(setup_00001_correct, "ModelString")
+        sim.simulate()
+        assert sim.results.status_list == [2]
+        sim.simulate(max_t=20.0, max_iter=5)
+        assert sim.results.status_list == [1]
+
+    def test_correct_model_file(self):
+        sim = Simulation.load_model("tests/models/00001.txt", "ModelFile")
+        sim.simulate()
+        assert sim.results.status_list == [2]